V47 - Complete text correct code, add safe file close on quit

+++ src/dispOCRtool/ocr/TextCorrection.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from PySide6.QtCore import Slot
 from PySide6.QtCore import QObject, QThread, QUrl, QElapsedTimer
 import language_tool_python
+from language_tool_python.utils import correct
 
 
 class TextCorrector(QThread):
@@ -19,21 +20,35 @@ class TextCorrector(QThread):
         self.lt = lt_instance if lt_instance is not None else language_tool_python.LanguageTool("en-US")
 
         self.timer = QElapsedTimer()  # For performance measure
-
+        self.aborted = False
 
     def run(self):
         src_path = Path(self.out_dir) / (self.filename + ".txt")
         dst_path = Path(self.out_dir) / (self.filename + "_corrected.txt")
 
-        self.timer.start()
-        with open(src_path, "r", encoding="utf-8") as src_file:
-            lines = src_file.readlines()
+        if not src_path.exists():
+            print(f"{self.filename}.txt does not exist.")
 
-        with open(dst_path, "w", encoding="utf-8") as dst_file:
-            for line in lines:
-                if line.strip():
-                    matches = self.lt.check(line)
-                    line = self.lt.utils.correct(line, matches)
+        else:
+            self.timer.start()
+            with open(src_path, "r", encoding="utf-8") as src_file:
+                lines = src_file.readlines()
 
-                dst_file.write(line)
-                print(self.timer.restart())
+            with open(dst_path, "w", encoding="utf-8") as dst_file:
+                for line in lines:
+                    print(line)
+                    if self.aborted: break
+                    if line.strip():
+                        matches = self.lt.check(line)
+                        line = correct(line, matches)
+
+                    dst_file.write(line)
+                    print(self.timer.restart())
+
+        print("> Text correct thread finished successfully.")
+
+    def stop(self):
+        self.aborted = True
+        self.requestInterruption()
+        self.wait()
+        print("> Text correct thread aborted successfully.")
